fix: move to freyja
