+++ packages/shared/src/features/onboarding/types/funnel.ts
@@ -338,10 +338,19 @@ export interface FunnelStepPlusCards
   }>;
 }
 
+export type FunnelExtensionImage = { default: string; retina: string };
+
 export interface FunnelStepBrowserExtension
   extends FunnelStepCommon<{
     headline: string;
     explainer: string;
+    cta: string;
+    skip: string;
+    showReviews: boolean;
+    image?: {
+      chrome?: FunnelExtensionImage;
+      edge?: FunnelExtensionImage;
+    };
   }> {
   type: FunnelStepType.BrowserExtension;
   onTransition: FunnelStepTransitionCallback<{
